+++ packages/terra-breakpoints/src/ActiveBreakpointProvider.jsx
@@ -23,10 +23,14 @@ class ActiveBreakpointProvider extends React.Component {
 
   componentDidMount() {
     window.addEventListener('resize', this.setActiveBreakpoint);
+
+    // This event is needed to handle mobile device rotation since the resize event is not triggered.
+    window.addEventListener('orientationchange', this.setActiveBreakpoint);
   }
 
   componentWillUnmount() {
     window.removeEventListener('resize', this.setActiveBreakpoint);
+    window.removeEventListener('orientationchange', this.setActiveBreakpoint);
   }
 
   setActiveBreakpoint() {