+++ packages/terra-dropdown-button/src/DropdownButton.jsx
@@ -1,13 +1,14 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import classNames from 'classnames/bind';
+import classNamesBind from 'classnames/bind';
+import ThemeContext from 'terra-theme-context';
 import * as KeyCode from 'keycode-js';
 import DropdownButtonBase from './_DropdownButtonBase';
 import styles from './DropdownButton.module.scss';
 import Item from './Item';
 import SplitButton, { Variants as SplitButtonVariants } from './SplitButton';
 
-const cx = classNames.bind(styles);
+const cx = classNamesBind.bind(styles);
 
 const Variants = {
   NEUTRAL: 'neutral',
@@ -141,6 +142,8 @@ class DropdownButton extends React.Component {
       ...customProps
     } = this.props;
 
+    const theme = this.context;
+
     const { isOpen, isActive, openedViaKeyboard } = this.state;
 
     const classnames = cx(
@@ -153,6 +156,7 @@ class DropdownButton extends React.Component {
         the dropdown open will cause the dropdown to close and reopen
       */
       { 'ignore-react-onclickoutside': isOpen },
+      theme.className,
     );
 
     return (
@@ -191,6 +195,7 @@ class DropdownButton extends React.Component {
 
 DropdownButton.propTypes = propTypes;
 DropdownButton.defaultProps = defaultProps;
+DropdownButton.contextType = ThemeContext;
 
 export default DropdownButton;
 export {