+++ packages/terra-form-select/src/tag/Menu.jsx
@@ -1,6 +1,7 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import classNames from 'classnames/bind';
+import classNamesBind from 'classnames/bind';
+import ThemeContext from 'terra-theme-context';
 import { polyfill } from 'react-lifecycles-compat';
 import { injectIntl, intlShape } from 'react-intl';
 import * as KeyCode from 'keycode-js';
@@ -9,7 +10,7 @@ import MaxSelection from '../shared/_MaxSelection';
 import MenuUtil from '../shared/_MenuUtil';
 import styles from '../shared/_Menu.module.scss';
 
-const cx = classNames.bind(styles);
+const cx = classNamesBind.bind(styles);
 
 /* Disabling this rule because props used in getDerivedStateFromProps return false positives. */
 /* eslint-disable react/no-unused-prop-types */
@@ -429,6 +430,8 @@ class Menu extends React.Component {
   }
 
   render() {
+    const theme = this.context;
+
     return (
       /**
        * Note: role="listbox" and aria-activedescendant needed for VoiceOver on iOS to properly
@@ -440,7 +443,7 @@ class Menu extends React.Component {
       <ul
         id="terra-select-menu"
         role="listbox"
-        className={cx('menu')}
+        className={cx('menu', theme.className)}
         aria-label={this.props.intl.formatMessage({ id: 'Terra.form.select.menu' })}
         ref={(menu) => { this.menu = menu; }}
         {...(this.state.active !== null ? { 'aria-activedescendant': `terra-select-option-${this.state.active}` } : {})}
@@ -454,6 +457,7 @@ class Menu extends React.Component {
 
 Menu.propTypes = propTypes;
 Menu.defaultProps = defaultProps;
+Menu.contextType = ThemeContext;
 
 /**
  * This polyfill enables backwards compatibility of features added in React 16.3.0.