+++ packages/terra-avatar/CHANGELOG.md
@@ -3,6 +3,11 @@ ChangeLog
 
 Unreleased
 ----------
+### Fixed
+* Updated jest snapshots for terra-image change.
+
+3.3.0 - (May 19, 2020)
+------------------
 ### Changed
 * Apply themes by context and directly include packaged themes.