+++ packages/terra-breakpoints/CHANGELOG.md
@@ -4,6 +4,11 @@ ChangeLog
 Unreleased
 ----------
 
+2.27.0 - (May 26, 2020)
+------------------
+### Fixed
+* Handle the `orientationchange` event on mobile device rotation since the `resize` event is not triggered.
+
 2.26.0 - (April 28, 2020)
 ------------------
 ### Changed