+++ packages/terra-demographics-banner/CHANGELOG.md
@@ -3,6 +3,14 @@ ChangeLog
 
 Unreleased
 ----------
+
+3.38.0 - (May 26, 2020)
+------------------
+### Changed
+* Minor dependency version bump
+
+3.37.0 - (May 19, 2020)
+------------------
 ### Changed
 * Apply themes by context and directly include packaged themes.