+++ packages/terra-form-input/CHANGELOG.md
@@ -4,6 +4,11 @@ ChangeLog
 Unreleased
 ----------
 
+3.4.0 - (May 19, 2020)
+------------------
+### Changes
+* Update wdio tests to move cursor to fix inconsistent hover styles in internal themes
+
 3.3.0 - (May 12, 2020)
 ------------------
 ### Changed