+++ packages/terra-form-textarea/CHANGELOG.md
@@ -4,6 +4,16 @@ ChangeLog
 Unreleased
 ----------
 
+4.5.0 - (May 26, 2020)
+------------------
+### Changed
+* Apply themes by context and directly include packaged themes.
+
+4.4.0 - (May 19, 2020)
+------------------
+### Changes
+* Update wdio tests to move cursor to fix inconsistent hover styles in internal themes
+
 4.3.0 - (May 12, 2020)
 ------------------
 ### Changed