+++ packages/terra-tag/CHANGELOG.md
@@ -4,6 +4,16 @@ ChangeLog
 Unreleased
 ----------
 
+2.34.0 - (May 26, 2020)
+------------------
+### Changed
+* Corrected typo in 'Terra-Tag' example doc.
+
+2.33.0 - (May 19, 2020)
+------------------
+### Changed
+* Minor dependency version bump
+
 2.32.0 - (April 28, 2020)
 ------------------
 ### Changed