+++ packages/terra-toggle-section-header/CHANGELOG.md
@@ -6,6 +6,11 @@ Unreleased
 ### Added
 * Added `isOpen` prop for programmatic toggle.
 
+2.35.0 - (May 19, 2020)
+------------------
+### Changed
+* Minor dependency version bump
+
 2.34.0 - (April 28, 2020)
 ------------------
 ### Changed